requirements.txt
numpy==1.24.3
opencv_python==4.8.0.74
scikit_image==0.21.0
scikit_learn==1.3.0
Pillow==10.0.0
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.100
